+++ internal/bin/cldev
@@ -41,14 +41,14 @@ case "$1" in
     ;;
   migration)
     timestamp=$(date +%s)
-    pkg="store/models/migrations/migration${timestamp}"
+    pkg="store/migrations/migration${timestamp}"
     echo Creating migration package $pkg
     mkdir -p $pkg
     echo package migration${timestamp}_test > $pkg/migrate_test.go
     cat > $pkg/migrate.go <<-EOF
 package migration$timestamp
 
-import "github.com/smartcontractkit/chainlink/store/models/orm"
+import "github.com/smartcontractkit/chainlink/store/orm"
 
 type Migration struct{}
